pyproject.toml
[tool.poetry]
name = "your-project-name"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]
packages = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
dspy-ai = "2.4.9"
wikipedia = "1.4.0"
sentence-transformers = "*"
toml = "*"
langchain-text-splitters = "*"
trafilatura = "*"
langchain-huggingface = "*"
qdrant-client = "*"
langchain-qdrant = "*"
numpy = "1.26.4"
streamlit = "1.31.1"
streamlit-card = "*"
markdown = "*"
unidecode = "*"
extra-streamlit-components = "0.1.60"
streamlit-extras = "*"
deprecation = "2.1.0"
st-pages = "0.4.5"
streamlit-float = "*"
streamlit-option-menu = "*"
python-fasthtml = "^0.10.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"